build(local): Improve buildout config layering
I was running into many problems re-using this local development support getting the
wrong set of buildout `./*.cfg` symlinks, overwriting the actual files in the VCS
checkout source, etc..  Now that I understand more I'm better able to define what we're
doing here so this change reworks it with that understanding in mind.  This should be
much more stable.