fixup! RELEASE build(profiles): Automatic upgrade all GS profiles
Use `collective.recipe.plonesite` and `collective.upgrade` to automatically apply all
upgrade steps, including for the Plone portal itself, when buildout is run.  Disabled by
default, for example when building the container image.  Enable by running with the CLI
option:

```
$ ./bin/buildout plonesite:enabled=True
```

I've used this approach in the past and once set up and integrated into the deployment
workflow it greatly simplifies the human workflow and thus also increases reliability.
That said anything that touches the deployment pipeline like this introduces significant
risks just due to the fact that unanticipated effects or concerns could impact any part
of the application or all of it.  So best to give this a thorough review: as many
reviewers as can be bothered and in as many environments as possible.

The basic idea is that a corresponding change is made to the the site-specific buildout
configuration that extends the base image buildout configuration changes in this commit.
Specifically, declare which GenericSetup profile is the "cannonical" policy profile for
the given site in the `profiles-initial` option of the `plonesite` buildout part.  For
`eea.docker.plone-eea-www` for example, this would be added to `./src/plone/www.cfg`:

    profiles-initial = Products.EEAPloneAdmin:default

Then when building is run with the `enabled` option as above, all profiles will be
upgraded including the Plone portal itself.  Additionally, when run against an empty
ZODB, such as when setting up a development environment without data, then that
"cannonical" policy profile will be installed automatically. Finally, when developing
upgrade steps the developer can test them with a higher degree of confidence that
they're testing it the way it will be deployed.

In our, containerized deployment environments we can't (and wouldn't want to)
run/install the `plonesite` part when buildout is first run, during the container image
build, so it's disabled by default.  Then it can be explicitly enabled to run at the
time services are started using a run-once `plonesite` service in
`./docker-compose.yml`.  A PR is forthcoming to do this for a specific site in
`eea.rancher.catalog` which also will serve as a demonstration of how we'd integrate
this into our deployment pipeline.

There are other ways to use `collective.recipe.plonesite`, including for deployment, so
please bring up anything that feels off as there may be easy adjustments that may
address those concerns.

P.S. We have to just live with the `plonesite` name.  I think it's awful and
communicates nothing or the wrong thing to someone who reads it, but there's a ton of
history behind it at this point.